% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sge_run.R
\name{sgePrepareScenarios}
\alias{sgePrepareScenarios}
\title{Run preparations for SGE submission}
\usage{
sgePrepareScenarios(
  expName,
  scenarios,
  bSize = 200,
  memCPU = "1G",
  nCPU = 1,
  time = "00:10:00",
  qos = "all.q",
  ...
)
}
\arguments{
\item{expName}{Name of experiment}

\item{scenarios}{Scenario data frame}

\item{bSize}{Number of parrellel processed files}

\item{memCPU}{Memory per CPU}

\item{nCPU}{Number of cores to reserve}

\item{time}{Maximum time}

\item{qos}{Quality of service}
}
\description{
Run preparations for SGE submission
}
