% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_scenarios.R
\name{slurmCreateScenarios}
\alias{slurmCreateScenarios}
\title{Submit scenario creation job to SLURM}
\usage{
slurmCreateScenarios(moreArgs = NULL, n = 0, limit = 3, ...)
}
\arguments{
\item{moreArgs}{Further arguments for SLURM's sbatch command, e.g.
c("--arg1", "--arg2")}

\item{n}{Current try.}

\item{limit}{Maximum number of tries/resubmissions.}

\item{...}{Deprecated arguments}
}
\description{
Submit scenario creation job to SLURM
}
