% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_results.R
\name{slurmPrepareResults}
\alias{slurmPrepareResults}
\title{Run preparations for SLURM submission}
\usage{
slurmPrepareResults(
  expDir,
  dbName,
  dbDir = NULL,
  resultsName = "results",
  resultsCols = c("scenario_id", "date_aggregation", "date", "age_group",
    "incidenceRate", "prevalenceRate"),
  indexOn = list(c("results", "scenario_id")),
  ncoresDT = 1,
  strategy = "batch",
  appendResults = FALSE,
  aggrFun = CalcEpiOutputs,
  aggrFunArgs = list(aggregateByAgeGroup = c("0-5", "2-10", "0-100"), aggregateByDate =
    "year"),
  splitBy = 10000,
  verbose = FALSE,
  ntasks = 1,
  mem = "32G",
  nCPU = 1,
  time = "06:00:00",
  qos = "6hours"
)
}
\arguments{
\item{expDir}{Directory of experiment}

\item{dbName}{Name of the database to create}

\item{dbDir}{Directory of the database file. Defaults to the root directory.}

\item{resultsName}{Name of the database table to add the results to.}

\item{resultsCols}{Vector of column names.}

\item{indexOn}{Define which index to create. Needs to be a lis of the form
list(c(TABLE, COLUMN), c(TABLE, COLUMN), ...).}

\item{ncoresDT}{Number of data.table threads to use on each parallel
cluster.}

\item{strategy}{Defines how to process the files. "batch" means that all
files are read into a single data frame first, then the aggregation
funciton is applied to that data frame and the result is added to the
database. "serial" means that each individual file is processed with the
aggregation function and added to the database.}

\item{appendResults}{If TRUE, do not add metadata to the database and only
write results.}

\item{aggrFun}{A function for aggregating the output of readFun. First
argument needs to be the output data frame of readFun and it needs to
generate a data frame. The data frame should NOT contain an experiment_id
column as this is added automatically. The column names needs to match the
ones defined in resultsCols.}

\item{aggrFunArgs}{Arguments for aggrFun as a (named) list.}

\item{splitBy}{Split total number of scenarios into groups. Either a numeric
value or a column name from your scenarios.}

\item{verbose}{Toggle verbose output}

\item{ntasks}{Number of tasks per CPU}

\item{mem}{Total amount of RAM}

\item{nCPU}{Number of cores to reserve}

\item{time}{Maximum time}

\item{qos}{Quality of service}
}
\description{
Run preparations for SLURM submission
}
