% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_scenarios.R
\name{slurmPrepareScenarios}
\alias{slurmPrepareScenarios}
\title{Run preparations for SLURM submission}
\usage{
slurmPrepareScenarios(
  expName,
  scenarios,
  bSize = 200,
  ntasks = 1,
  memCPU = "1G",
  nCPU = 1,
  time = "00:10:00",
  qos = "30min"
)
}
\arguments{
\item{expName}{Name of experiment}

\item{scenarios}{Scenario data frame}

\item{bSize}{Number of parrellel processed files}

\item{ntasks}{Number of tasks per CPU}

\item{memCPU}{Memory per CPU}

\item{nCPU}{Number of cores to reserve}

\item{time}{Maximum time}

\item{qos}{Quality of service}
}
\description{
Run preparations for SLURM submission
}
