% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_simulation.R
\name{slurmPrepareSimulations}
\alias{slurmPrepareSimulations}
\title{Run preparations for SLURM submission}
\usage{
slurmPrepareSimulations(
  expName,
  scenarios,
  bSize = 200,
  ntasks = 1,
  memCPU = "1G",
  nCPU = 1,
  time = "06:00:00",
  qos = "6hours",
  verbose = FALSE
)
}
\arguments{
\item{expName}{Name of experiment}

\item{scenarios}{Scenario data frame}

\item{bSize}{Number of parrellel processed files}

\item{ntasks}{Number of tasks per CPU}

\item{memCPU}{Memory per CPU}

\item{nCPU}{Number of cores to reserve}

\item{time}{Maximum time}

\item{qos}{Quality of service}

\item{verbose}{If TRUE, use OpenMalaria's verbose output.}
}
\description{
Run preparations for SLURM submission
}
