% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_results.R
\name{slurmRunResults}
\alias{slurmRunResults}
\title{Submit results job to SLURM}
\usage{
slurmRunResults(moreArgs = NULL, ...)
}
\arguments{
\item{moreArgs}{Further arguments for SLURM's sbatch command, e.g.
c("--arg1", "--arg2")}

\item{...}{Deprecated arguments}
}
\description{
Submit results job to SLURM
}
