% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_simulation.R
\name{slurmRunSimulation}
\alias{slurmRunSimulation}
\title{Submit simulation job to SLURM}
\usage{
slurmRunSimulation(moreArgs = NULL, n = 0, limit = 3, ...)
}
\arguments{
\item{moreArgs}{Further arguments for SLURM's sbatch command, e.g.
c("--arg1", "--arg2")}

\item{n}{Current try.}

\item{limit}{Maximum number of tries/resubmissions.}

\item{...}{Deprecated arguments}
}
\description{
Submit simulation job to SLURM
}
